move campaign_dates stuff to diagnostic_teledeclaration_dates
