+++ api/serializers/teledeclaration.py
@@ -13,19 +13,3 @@ class ShortTeledeclarationSerializer(serializers.ModelSerializer):
             "status",
         )
         read_only_fields = fields
-
-
-class CampaignDatesSerializer(serializers.Serializer):
-    year = serializers.IntegerField()
-    teledeclaration_start_date = serializers.DateTimeField()
-    teledeclaration_end_date = serializers.DateTimeField()
-    correction_start_date = serializers.DateTimeField(allow_null=True)
-    correction_end_date = serializers.DateTimeField(allow_null=True)
-    legifrance_url = serializers.URLField(allow_null=True)
-    rapport_parlement_url = serializers.URLField(allow_null=True)
-
-
-class CampaignDatesFullSerializer(CampaignDatesSerializer):
-    # additional fields
-    in_teledeclaration = serializers.BooleanField()
-    in_correction = serializers.BooleanField()
