+++ api/views/__init__.py
@@ -83,7 +83,7 @@ from .review import ReviewView  # noqa: F401
 from .sector import SectorListView  # noqa: F401
 from .statistics import CanteenStatisticsView  # noqa: F401
 from .subscription import SubscribeNewsletter  # noqa: F401
-from .teledeclaration import (  # noqa: F401
+from .diagnostic_teledeclaration_dates import (  # noqa: F401
     TeledeclarationCampaignDatesListView,
     TeledeclarationCampaignDatesRetrieveView,
 )
